test(appeals): updating api command names and creating new command files
