+++ appeals/e2e/cypress/e2e/back-office-appeals/manageDocsAppellant.spec.js
@@ -34,7 +34,7 @@ describe('manage docs on appellant case', () => {
 
 	const setupInquiry = (caseObj, inquiryDate) => {
 		// require case to be started as inquiry to access appellant POE evidence e.g.
-		cy.addInquiryViaApi(caseObj, inquiryDate);
+		cy.addInquiry(caseObj, inquiryDate);
 		happyPathHelper.assignCaseOfficer(caseObj);
 		happyPathHelper.reviewAppellantCase(caseObj);
 		happyPathHelper.startS78InquiryCase(caseObj, 'inquiry');
