+++ appeals/e2e/cypress/e2e/back-office-test-data-generation/createHouseholderAppeal.spec.js
@@ -37,13 +37,13 @@ describe('Create householder application', () => {
 
 			// Assign Case Officer Via API
 			if (testCaseConfig.assignCaseOfficer) {
-				cy.assignCaseOfficerViaApi(caseObj);
+				cy.assignCaseOfficer(caseObj);
 			}
 
 			// Validate Appeal Via API
 			if (testCaseConfig.validateAppeal) {
 				cy.getBusinessActualDate(new Date(), 0).then((date) => {
-					cy.updateAppealDetailsViaApi(caseObj, { validationOutcome: 'valid', validAt: date });
+					cy.updateAppealDetails(caseObj, { validationOutcome: 'valid', validAt: date });
 				});
 			}
 
@@ -51,7 +51,7 @@ describe('Create householder application', () => {
 			if (testCaseConfig.addInquiry) {
 				inquirySectionPage.setupTimetableDates().then(({ currentDate, ...timeTable }) => {
 					// Set up inquiry case with LPA questionnaire
-					cy.addInquiryViaApi(caseObj, currentDate, timeTable);
+					cy.addInquiry(caseObj, currentDate, timeTable);
 				});
 			}
 
@@ -61,7 +61,7 @@ describe('Create householder application', () => {
 					// Set up hearing case
 					cy.addHearingDetails(caseObj, date);
 				});
-				cy.setupHearingViaApi(caseObj);
+				cy.setupHearing(caseObj);
 			}
 
 			// add lpaq
